+++ zcash_client_sqlite/src/wallet/transparent.rs
@@ -1307,21 +1307,13 @@ pub(crate) fn queue_transparent_spend_detection<P: consensus::Parameters>(
 
 #[cfg(test)]
 mod tests {
-    use std::collections::BTreeMap;
-
     use secrecy::Secret;
     use transparent::keys::NonHardenedChildIndex;
     use zcash_client_backend::{
-        data_api::{
-            testing::{AddressType, TestBuilder},
-            wallet::decrypt_and_store_transaction,
-            Account as _, WalletRead, WalletWrite,
-        },
+        data_api::{testing::TestBuilder, Account as _, WalletWrite},
         wallet::TransparentAddressMetadata,
     };
-    use zcash_keys::address::Address;
     use zcash_primitives::block::BlockHash;
-    use zcash_protocol::value::Zatoshis;
 
     use crate::{
         error::SqliteClientError,
@@ -1358,120 +1350,12 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "transparent-inputs")]
     fn gap_limits() {
-        let mut st = TestBuilder::new()
-            .with_data_store_factory(TestDbFactory::default())
-            .with_block_cache(BlockCache::new())
-            .with_account_from_sapling_activation(BlockHash([0; 32]))
-            .build();
-
-        let test_account = st.test_account().cloned().unwrap();
-        let account_uuid = test_account.account().id();
-        let ufvk = test_account.account().ufvk().unwrap().clone();
-
-        let external_taddrs = st
-            .wallet()
-            .get_transparent_receivers(account_uuid, false)
-            .unwrap();
-        assert_eq!(
-            u32::try_from(external_taddrs.len()).unwrap(),
-            GapLimits::default().external()
+        zcash_client_backend::data_api::testing::transparent::gap_limits(
+            TestDbFactory::default(),
+            BlockCache::new(),
+            GapLimits::default().into(),
         );
-        let internal_taddrs = st
-            .wallet()
-            .get_transparent_receivers(account_uuid, true)
-            .unwrap();
-        assert_eq!(
-            u32::try_from(internal_taddrs.len()).unwrap(),
-            GapLimits::default().external() + GapLimits::default().internal()
-        );
-        let ephemeral_taddrs = st
-            .wallet()
-            .get_known_ephemeral_addresses(account_uuid, None)
-            .unwrap();
-        assert_eq!(
-            u32::try_from(ephemeral_taddrs.len()).unwrap(),
-            GapLimits::default().ephemeral()
-        );
-
-        // Add some funds to the wallet
-        let (h0, _, _) = st.generate_next_block(
-            &ufvk.sapling().unwrap(),
-            AddressType::DefaultExternal,
-            Zatoshis::const_from_u64(1000000),
-        );
-        st.scan_cached_blocks(h0, 1);
-
-        // The previous operation was shielded-only, but unified address usage within the
-        // valid transparent child index range still count towards the gap limit, so this
-        // updates the gap limit by the index of the default Sapling receiver
-        let external_taddrs = st
-            .wallet()
-            .get_transparent_receivers(account_uuid, false)
-            .unwrap();
-        assert_eq!(
-            u32::try_from(external_taddrs.len()).unwrap(),
-            GapLimits::default().external()
-                + (u32::try_from(ufvk.sapling().unwrap().default_address().0).unwrap() + 1)
-        );
-
-        // Pick an address half way through the set of external taddrs
-        let external_taddrs_sorted = external_taddrs
-            .into_iter()
-            .filter_map(|(addr, meta)| meta.map(|m| (m.address_index(), addr)))
-            .collect::<BTreeMap<_, _>>();
-        let to = Address::from(
-            *external_taddrs_sorted
-                .get(&NonHardenedChildIndex::from_index(4).unwrap())
-                .expect("An address exists at index 4."),
-        )
-        .to_zcash_address(st.network());
-
-        // Create a transaction & scan the block. Since the txid corresponds to one our wallet
-        // generated, this should cause the gap limit to be bumped (generating addresses with index
-        // 10..15
-        let txids = st
-            .create_standard_transaction(&test_account, to, Zatoshis::const_from_u64(20000))
-            .unwrap();
-        let (h1, _) = st.generate_next_block_including(txids.head);
-
-        // At this point, the transaction has been created, but since it has not been mined it does
-        // not cause an update to the gap limit; we have to wait for the transaction to actually be
-        // mined or we could bump the gap limit too soon and start generating addresses that will
-        // never be inspected on wallet recovery.
-        let external_taddrs = st
-            .wallet()
-            .get_transparent_receivers(account_uuid, false)
-            .unwrap();
-        assert_eq!(
-            u32::try_from(external_taddrs.len()).unwrap(),
-            GapLimits::default().external()
-                + (u32::try_from(ufvk.sapling().unwrap().default_address().0).unwrap() + 1)
-        );
-
-        // Mine the block, then use `decrypt_and_store_transaction` to ensure that the wallet sees
-        // the transaction as mined (since transparent handling doesn't get this from
-        // `scan_cached_blocks`)
-        st.scan_cached_blocks(h1, 1);
-        let tx = st.wallet().get_transaction(txids.head).unwrap().unwrap();
-        decrypt_and_store_transaction(&st.network().clone(), st.wallet_mut(), &tx, Some(h1))
-            .unwrap();
-
-        // Now that the transaction has been mined, the gap limit should have increased.
-        let external_taddrs = st
-            .wallet()
-            .get_transparent_receivers(account_uuid, false)
-            .unwrap();
-        assert_eq!(
-            u32::try_from(external_taddrs.len()).unwrap(),
-            GapLimits::default().external() + 5
-        );
-
-        // The utxo query height should be equal to the minimum mined height among transactions
-        // sent to any of the set of {addresses in the gap limit range | address prior to the gap}.
-        let query_height = st.wallet().utxo_query_height(account_uuid).unwrap();
-        assert_eq!(query_height, h0);
     }
 
     #[test]